+++ src/ui/toolbar/Toolbar.tsx
@@ -10,7 +10,7 @@ import { useUpdater } from "./useUpdater";
 import UpdaterIcon from "./UpdaterIcon";
 import UpdaterModal from "./UpdaterModal";
 import favicon from "./favicon.svg";
-import QuickSyncSection from "./QuickSyncSection";
+
 
 export default function Toolbar() {
     const project = useContext(ProjectContext)
@@ -40,7 +40,6 @@ export default function Toolbar() {
                         <NavbarHeading>
                             Project: <strong>{project.name}</strong> {project.dirty && "(unsaved)"}
                         </NavbarHeading>
-                        <QuickSyncSection />
                     </>
                 }
             </NavbarGroup>