+++ src/ui/workarea/canvas/KerningPreview.tsx
@@ -6,17 +6,19 @@ import { calculateTemplateData, generateTemplatePng } from "../../../generation/
 import { Card } from "@blueprintjs/core"
 import useResizeObserver from "@react-hook/resize-observer"
 import { useProjectState } from "../hooks/useProjectState"
+import { useIncrementingCounter } from "../hooks/useIncrementingCounter"
 
 export interface Props {
     project: ProjectData
-    kerning: KerningPair
+    kernings: KerningPair[]
 }
 
-export default function KerningPreview({ project, kerning }: Props) {
+export default function KerningPreview({ project, kernings }: Props) {
     const [scale] = useProjectState("previewScale", project)
     const [bgColor] = useProjectState("previewBgColor", project)
     const [ctx, setCtx] = useState<CanvasRenderingContext2D | null>(null)
     const [font, setFont] = useState<[FontSpec, Blob[]] | null>(null)
+    const previewCounter = useIncrementingCounter(1, 1000)
 
     const canvasParentRef = useRef<HTMLDivElement | null>(null)
     const canvasRef = useRef<HTMLCanvasElement | null>(null)
@@ -32,31 +34,35 @@ export default function KerningPreview({ project, kerning }: Props) {
     useEffect(() => {
         const generate = async () => {
             const templateData = calculateTemplateData(project, "current or imported")
-            templateData.project.kernings = [kerning]
+            templateData.project.kernings = [...kernings]
             const templateImage = await generateTemplatePng(templateData)
             const font = await generateFont(templateData, templateImage)
             
             setFont(font)
         }
         generate()
-    }, [project, kerning])
+    }, [project, kernings])
 
     useEffect(() => {
         if (ctx && font) {
             const draw = async () => {
                 const [spec, pages] = font
-                drawPreview(`${String.fromCharCode(kerning.first)}${String.fromCharCode(kerning.second)}`, spec, pages, scale, ctx)
+                const leftChar = String.fromCodePoint(kernings[previewCounter % kernings.length]?.first ?? 0)
+                const rightChar = String.fromCodePoint(kernings[previewCounter % kernings.length]?.second ?? 0)
+                drawPreview(`${leftChar}${rightChar}`, spec, pages, scale, ctx)
             }
             draw()
         }
-    }, [kerning, kerning, ctx, font])
+    }, [kernings, ctx, font, previewCounter])
 
     useResizeObserver(canvasParentRef, entry => {
         if (canvasRef.current) {
             canvasRef.current.width = entry.contentRect.width
             if (ctx && font) {
                 const [spec, pages] = font
-                drawPreview(`${String.fromCharCode(kerning.first)}${String.fromCharCode(kerning.second)}`, spec, pages, scale, ctx)
+                const leftChar = String.fromCodePoint(kernings[previewCounter % kernings.length]?.first ?? 0)
+                const rightChar = String.fromCodePoint(kernings[previewCounter % kernings.length]?.second ?? 0)
+                drawPreview(`${leftChar}${rightChar}`, spec, pages, scale, ctx)
             }
         }
     })