+++ src/ui/workarea/forms/PreviewTemplateButton.tsx
@@ -2,6 +2,8 @@ import { useState } from "react";
 import { ProjectData } from "../../../filesystem/projectstore";
 import { Button, FormGroup } from "@blueprintjs/core";
 import PreviewTemplateDialog from "./dialogs/PreviewTemplateDialog";
+import ExportTemplateDialog from "../../../ui/toolbar/dialogs/ExportTemplateDialog";
+import styles from "../workarea.module.scss";
 
 export interface Props {
     project: ProjectData
@@ -9,18 +11,27 @@ export interface Props {
 }
 
 export default function PreviewTemplateButton({ project, forceDisabled }: Props) {
-    const [dialogOpen, setDialogOpen] = useState(false)
+    const [previewOpen, setPreviewOpen] = useState(false)
+    const [exportOpen, setExportOpen] = useState(false)
 
     return (
         <>
             <FormGroup>
-                <Button icon="eye-open" text="Preview Template" onClick={() => setDialogOpen(true)} disabled={forceDisabled} />
+                <div className={styles.buttonrow}>
+                    <Button icon="eye-open" text="Preview Template" onClick={() => setPreviewOpen(true)} disabled={forceDisabled} size="large" className={styles.button} />
+                    <Button icon="export" text="Export Template" onClick={() => setExportOpen(true)} disabled={forceDisabled} size="large" className={styles.button} />
+                </div>
             </FormGroup>
 
             <PreviewTemplateDialog
                 project={project}
-                isOpen={dialogOpen}
-                setIsOpen={setDialogOpen}
+                isOpen={previewOpen}
+                setIsOpen={setPreviewOpen}
+            />
+            <ExportTemplateDialog
+                project={project}
+                isOpen={exportOpen}
+                setIsOpen={setExportOpen}
             />
         </>
     )