+++ src/ui/workarea/forms/dialogs/AddKerningDialog.tsx
@@ -1,5 +1,5 @@
 import { Button, Callout, Dialog, DialogBody, DialogFooter, FormGroup, InputGroup } from "@blueprintjs/core"
-import { ProjectData } from "../../../../filesystem/projectstore"
+import { KerningPair, ProjectData } from "../../../../filesystem/projectstore"
 import KerningPreview from "../../canvas/KerningPreview"
 import { useContext, useState } from "react"
 import { ProjectMutContext } from "../../../contexts/ProjectContext"
@@ -15,8 +15,8 @@ export default function AddKerningDialog({
     isOpen,
     setIsOpen,
 }: Props) {
-    const [left, setLeft] = useState<number | null>(null)
-    const [right, setRight] = useState<number | null>(null)
+    const [left, setLeft] = useState<string | null>(null)
+    const [right, setRight] = useState<string | null>(null)
     const [amount, setAmount] = useState("0")
     const [errorMessage, setErrorMessage] = useState<string | null>(null)
 
@@ -29,6 +29,19 @@ export default function AddKerningDialog({
         setRight(null)
         setAmount("0")
     }
+    
+    const willOverwrite = project.kernings.some(kerning => left?.includes(String.fromCodePoint(kerning.first)) && right?.includes(String.fromCodePoint(kerning.second)))
+    const permutations: KerningPair[] = [...(left ?? String.fromCodePoint(0))]
+        .flatMap(leftChar => [...(right ?? String.fromCodePoint(0))]
+            .map(rightChar => ({
+                first: leftChar.codePointAt(0) ?? 0,
+                second: rightChar.codePointAt(0) ?? 0,
+                amount: Number(amount),
+            } satisfies KerningPair))
+        )
+    const uniquePermutations = permutations.filter((perm, index, self) =>index === self.findIndex(p => p.first === perm.first && p.second === perm.second))
+    const invalidChar = left && [...left].find(char => !project.characterSet.includes(char))
+        || right && [...right].find(char => !project.characterSet.includes(char))
 
     const onAdd = () => {
         try {
@@ -36,18 +49,9 @@ export default function AddKerningDialog({
                 throw new Error("Characters aren't specified")
             }
 
-            project.kernings.forEach(kerning => {
-                if (kerning.first === left && kerning.second === right) {
-                    throw new Error("Kerning pair already exists")
-                }
-            })
-
             const updatedProject = { ...project, dirty: true }
-            updatedProject.kernings.push({
-                first: left,
-                second: right,
-                amount: Number(amount),
-            })
+            updatedProject.kernings = updatedProject.kernings.filter(kerning => !uniquePermutations.some(k => k.first === kerning.first && k.second === kerning.second))
+            updatedProject.kernings.push(...uniquePermutations)
             setProjectData(updatedProject)
             onClose()
         } catch (e: any) {
@@ -63,23 +67,20 @@ export default function AddKerningDialog({
             onClose={onClose}
         >
             <DialogBody>
-                <Callout>
-                    Overwrites how much closer or further apart should two characters be when next to each other.
-                    <br />
-                    <br />
-                    Not all game engines support this feature.
+                <Callout icon="info-sign">
+                    Input multiple characters to create all permutations of kerning pairs
                 </Callout>
                 <br />
-                <FormGroup label="Left character">
+                <FormGroup label="Left characters">
                     <InputGroup
-                        value={left === null ? "" : String.fromCharCode(left)}
-                        onValueChange={value => setLeft(value.charCodeAt(value.length-1))}
+                        value={left ?? ""}
+                        onValueChange={value => setLeft(value)}
                     />
                 </FormGroup>
-                <FormGroup label="Right character">
+                <FormGroup label="Right characters">
                     <InputGroup
-                         value={right === null ? "" : String.fromCharCode(right)}
-                         onValueChange={value => setRight(value.charCodeAt(value.length-1))}
+                         value={right ?? ""}
+                         onValueChange={value => setRight(value)}
                     />
                 </FormGroup>
                 <FormGroup label="Amount">
@@ -89,15 +90,20 @@ export default function AddKerningDialog({
                         type="number"
                     />
                 </FormGroup>
-                <KerningPreview project={project} kerning={{
-                    first: left ?? 0,
-                    second: right ?? 0,
-                    amount: Number(amount),
-                }} />
-                { ((left !== null && !project.characterSet.includes(String.fromCharCode(left)))
-                    || (right !== null && !project.characterSet.includes(String.fromCharCode(right)))) &&
+                <KerningPreview project={project} kernings={permutations} />
+                { invalidChar &&
+                    <Callout intent="danger">
+                        Character '<strong>{invalidChar}</strong>' is not in the character set
+                    </Callout>
+                }
+                { willOverwrite &&
                     <Callout intent="warning">
-                        At least one character is not in the current character set
+                        One or more existing kerning pairs will be overwritten
+                    </Callout>
+                }
+                { uniquePermutations.length !== permutations.length &&
+                    <Callout intent="warning">
+                        Some permutations are duplicates and will be ignored ({permutations.length - uniquePermutations.length} duplicates)
                     </Callout>
                 }
                 { errorMessage &&
@@ -109,7 +115,12 @@ export default function AddKerningDialog({
             <DialogFooter actions={
                 <>
                     <Button text="Cancel" onClick={onClose} />
-                    <Button intent="primary" text="Add" onClick={onAdd} />
+                    <Button
+                        intent="primary"
+                        text={`Add ${uniquePermutations.length} Pair${uniquePermutations.length !== 1 ? "s" : ""}`} 
+                        onClick={onAdd} 
+                        disabled={!left || !right || !!invalidChar}
+                    />
                 </>
             } />
         </Dialog>